Divise funzioni su server

+++ SBOBINATOR/Converter/views.py
@@ -4,16 +4,15 @@ from django.shortcuts import redirect
 import speech_recognition as sr
 import moviepy.editor as mp
 from django.http import HttpResponse
+from django.http import FileResponse
 import os
 import json
 
 def redirect_view(request):
     return redirect('/static/index.html')
 
-# Create your views here.
-def upload(request):
+def videoToAudio(request):
     if request.method == 'POST':
-        response_data = {}
         d = 'media'
         fs = FileSystemStorage()
         try:
@@ -24,6 +23,24 @@ def upload(request):
             path = path+'.wav'
             video.audio.write_audiofile(path)
             fs.delete(path.replace('.wav',''))
+            response = FileResponse(open(path, 'rb'))
+            os.remove(path)
+            return response
+        except Exception as e:
+            if os._exists(d):
+                for f in os.listdir(d):
+                    os.remove(os.path.join(d,f))
+            return 'error'
+
+def audioToText(request):
+    if request.method == 'POST':
+        response_data = {}
+        d = 'media'
+        fs = FileSystemStorage()
+        try:
+            uploaded_file = request.FILES['document']
+            path = os.path.join(d, uploaded_file.name)
+            fs.save(path, uploaded_file)
             text = conversion(path)
             fs.delete(path)
             response_data['text'] = text
